Refactor check_stock modules
